==> 02_make_main_workshop_list.Rmd <==
---
title: "Create CSV file of Future Workshops"
subtitle: "Data are from the LibCal API, parsed by rvest and dplyr"
date: "`r Sys.Date()`"
---

1. Run this at the beginning of the semester to generate the main workshop list
1. Use this to generate the workshop flyer, for advertising
1. This script depends on `01_gather_and_wrangle.R` to harvest the data

## Release Notes

- August 10, 2021:  Undertook this revision because the integration of online and in-person workshops made Joel's script ineffective.  Nonetheless, borrowed heavily from his previous work.  The RMD migration of Joel's `rvestLibcalCode.R`

## library packages

```{r message=FALSE, warning=FALSE}
Sys.setenv(TZ="America/New_York")
library(fs)
library(tidyverse)
library(lubridate)
```

## Get Data

Ingest data from web page (LibCal API) & make a column headers template for the attendance sheets that will be written to Google.

### harvest webpage data from LibCal API

this next CODE CHUNK is DISABLED, but designed to rerun the harvest script.  Manually, you must click the green execute arrow.


```{r get wrangled data, include=FALSE}
source("01_gather_and_wrangle.R", local = knitr::knit_global())
# or sys.source("your-script.R", envir = knitr::knit_global())
```

## Wrangle Data

```{r}
for_goog_flyer_df <- my_df |> 
    mutate(description = str_trim(str_remove_all(description, "\\[(Online|In-person)\\]"))) |> 
    # select(Workshop = title, Date = day_flyer, Time = time, Mode = online_in_person_flyer, registration_link, description) 
    select(Workshop = title, Date = day_flyer, Time = time_flyer, Time_begins = begins_display, Mode = online_in_person_flyer, registration_link, description) 

for_goog_flyer_df 
```


## Write output



```{r}
fs::dir_create("output")

fs::file_copy("output/workshops.csv", 
              glue::glue("output/workshops.bak.{Sys.Date()}.{round(Sys.getpid() * runif(1))}.csv"))

write_csv(my_df, "output/workshops.csv")

write_csv(for_goog_flyer_df, "output/forgoog_flyer.csv")  
```


## Next steps

See "Manually upload"

### above chunk 

may have to be be run manually.  It will upload to your google drive home directory (Look in the "recent" folder on google drive).  Then move to the proper path in Google Drive.  **BE SURE TO CHECK THE DATES**.  Google drive has a tendency to foul-up the dates and impose GMT instead of EST!!   I noticed this when manually uploading.  See note...


### Manual upload notes

1. `output/forgoogflyer.csv` to:  `Google Drive > CDVS > Services > Workshops > [semester] > Planning`
    - rename `for_flyer_joel_eric`
2. `output/workshops.csv` to: `Google Drive > CDVS > Services > Workshops > [semester] > Planning`
    - Rename and prepend with DO_NOT_MODIFY_workshops_spring_2023.csv
    - check dates
3. Then open the output/workshops.csv copy select columns from excel, paste to notepad++; then  re-copy and paste into Google Drive (GD)
    - paste **FROM** notepad++ **INTO** the Google Drive sheet `DVS_Workshops_summary_stats` (`GD > CDVS > Workshops > _Assessment > longitudinal`) columns A, B, and C:  
        - Excel column B to notepad to GD column A     [date]
        - Excel column H to notepad to GD column B     [workshop id]
        - Excel column A to notepad to GD column C     [workshop name] 
    




#### Note  
- Fall 2024 required a lot of rewriting in the 01 and 02 scripts.  The SpringShare API changes often.
- Ran this during spring of 2023.  It all worked fine.  
  - By spring of 2022 this problem had resolved but in the fall of 2021 there were extraneous leading single apostrophes showing up in the uploaded data.  This was probably a bug with `library(googlesheets4)`. The manual work-around, instead of `gs4_create()`, was to upload the forgoog_flyer.csv file **manually** to Google Drive, then open the csv on Google Drive, and then "Open With" as a Sheet.  BUT, this allowed Google to force it's own GMT timezone.  So, you have to open blank sheet (sheets.new), Then `import`, then be sure not to allow any automatic conversion of fields (this is an option in the later stages of the import.)
